Refactor string formatting code to utils
